Add customer info, store logo, and estimated pickup date to receipts
- Add customer name, phone, and email to all receipt types when available
- Add store logo upload functionality with Filament form
- Update all receipt templates to use store logo (fallback to default)
- Add estimated pickup date support for deferred payments
- Make copy receipt identical to sales receipt except for KOPI marking
- Change delivery receipt label from 'Utleveringskvittering' to 'Bestillingsbekreftelse'
- Fix spelling: IKKJE -> IKKE on all receipt types
- Make estimated pickup date bold on delivery receipts
- Update FlutterFlow custom action to support estimated pickup date as DateTime parameter
- Add logo_path field to Store model and migration

+++ docs/flutterflow/custom-actions/add_item_to_cart.dart
@@ -2,14 +2,14 @@
 // 
 // Parameters:
 //   - product (Product type) - Required
-//   - quantity (Integer) - Default: 1
+//   - quantity (Double) - Default: 1.0 (supports decimals for continuous units)
 //   - variant (ProductVariant) - Optional
 //
 // This action adds an item to the shopping cart or updates quantity if item already exists
 
 void addItemToCart(
   dynamic product,
-  int quantity,
+  double quantity,
   dynamic variant,
 ) {
   // Get current cart from app state
@@ -51,7 +51,7 @@ void addItemToCart(
       productName: existingItem.productName,
       productImageUrl: existingItem.productImageUrl,
       unitPrice: existingItem.unitPrice,
-      quantity: existingItem.quantity + quantity,
+      quantity: existingItem.quantity + quantity, // Both are now double
       originalPrice: existingItem.originalPrice,
       discountAmount: existingItem.discountAmount,
       discountReason: existingItem.discountReason,
